Allow writing of large HTML contents
Replace DataOutputStream.writeUtf with WritableUtils.writeString.
This fixes problems when the HTML pages created by Twitter are larger
than the limit of DataOutputStream (32KB).

The HtmlStatus objects are written with VERSION number 2. The readFields
method uses the correct way of reading depending on the version number,
therefore this change should be backwards compatible.